[frontend] Rename types.ts to dtos.ts
