+++ frontend/app/ui/user/match-history.tsx
@@ -1,5 +1,5 @@
 import { getMatchHistory } from "@/app/lib/actions";
-import type { MatchDetailEntity } from "@/app/lib/types";
+import type { MatchDetailEntity } from "@/app/lib/dtos";
 import Link from "next/link";
 import { Avatar } from "./avatar";
 import ProfileItem from "./profile-item";
